Renamed 'utils' directory to 'tools'.
